requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
Automat==24.8.1
azure-core==1.32.0
azure-identity==1.21.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
contextlib2==21.6.0
contourpy==1.3.1
courlan==1.3.2
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
django-ninja==1.3.0
django-ninja-extra==0.22.6
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
gradio==4.39.0
gradio_client==1.1.1
greenlet==3.1.1
h11==0.14.0
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
hyperlink==21.0.0
idna==3.10
importlib_resources==6.5.2
incremental==24.7.2
injector==0.22.0
ipython==7.34.0
itemadapter==0.11.0
itemloaders==1.3.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jusText==3.0.2
kiwisolver==1.4.8
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langsmith==0.1.147
lavague==1.1.19
lavague-contexts-openai==0.2.4
lavague-core==0.2.35
lavague-drivers-selenium==0.2.15
lavague-gradio==0.2.8
llama-cloud==0.1.15
llama-index==0.10.56
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.56
llama-index-embeddings-azure-openai==0.1.11
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post4
llama-index-llms-azure-openai==0.1.10
llama-index-llms-openai==0.1.26
llama-index-multi-modal-llms-azure-openai==0.1.4
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-retrievers-bm25==0.1.5
llama-parse==0.4.9
lxml==5.3.1
lxml_html_clean==0.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
msal==1.32.0
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
openai==1.67.0
orjson==3.10.15
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
prompt_toolkit==3.0.50
propcache==0.3.0
Protego==0.4.0
proto-plus==1.26.1
protobuf==6.30.1
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypdf==4.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-whois==0.9.5
pytz==2025.1
PyYAML==6.0.2
queuelib==1.7.0
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
ruff==0.11.1
Scrapy==2.12.0
seaborn==0.13.2
selenium==4.29.0
semantic-version==2.10.0
service-identity==24.2.0
setuptools==77.0.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.39
sqlparse==0.5.3
starlette==0.46.1
striprtf==0.0.26
tenacity==8.3.0
tiktoken==0.9.0
tld==0.13
tldextract==5.1.3
tomlkit==0.12.0
tqdm==4.67.1
trafilatura==1.12.2
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
w3lib==2.3.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
